refactored to move knowledge of BanyPlugin to outermost layer
- introduced new interface that is used by instances of
  ThirdPartyTransactionGateway. Implementations wrap BanyPlugin
  instances
- refactored to use OffsetDateTime instead of LocalDateTime
- refactored CibcTransaction to use BigDecimal under the hood

+++ src/main/kotlin/com/pissiphany/bany/dataStructure/RetrofitAccount.kt
@@ -4,5 +4,9 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class RetrofitAccount(
-    val id: String, val name: String, val closed: Boolean, val balance: Long, val type: String
+    val id: String,
+    val name: String,
+    val closed: Boolean,
+    val balance: Int,
+    val type: String
 )
